backend/requirements.txt
google-generativeai
GitPython
fastapi
uvicorn
python-dotenv
python-multipart
httpx
pytest
